include

+++ netc/build_scene.h
@@ -1,10 +1,6 @@
 #ifndef _BUILD_SCENE_H_
 #define _BUILD_SCENE_H_
 
-#include <array>
-#include <unordered_set>
-#include <functional>
-
 #include "deck_data.h"
 #include "scene_mgr.h"
 #include "gui_extra.h"
@@ -17,10 +13,10 @@ namespace ygopro
     struct BuilderCard : public DeckCardExtraData {
         unsigned int buffer_index = 0;
         ti4 card_tex;
-        float moving_time_b = 0.0f;
-        float moving_time_e = 0.0f;
-        float fading_time_b = 0.0f;
-        float fading_time_e = 0.0f;
+        double moving_time_b = 0.0;
+        double moving_time_e = 0.0;
+        double fading_time_b = 0.0;
+        double fading_time_e = 0.0;
         v2f pos = {0.0f, 0.0f};
         v2f size = {0.0f, 0.0f};
         v2f start_pos = {0.0f, 0.0f};
@@ -128,7 +124,7 @@ namespace ygopro
         bool show_exclusive = true;
         bool show_info = false;
         bool show_info_begin = false;
-        float show_info_time = 0.0f;
+        double show_info_time = 0.0;
         std::array<ti4, 3> limit;
         std::array<ti4, 3> pool;
         ti4 hmask;
